pyproject.toml
[project]
name = "solace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.39.0",
    "langchain-memgraph>=0.1.12",
    "mem0ai[graph]>=1.0.7",
    "numpy>=1.26.0",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "sounddevice>=0.5.0",
    "websockets>=13.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "neo4j>=5.0.0",
    "asyncpg>=0.29.0",
]
